% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brewer.R
\name{brewer}
\alias{brewer}
\title{Brewer sampling procedure --------------------------------------------------}
\usage{
brewer(pik, n, N, s, list)
}
\arguments{
\item{pik}{vector of first-order inclusion probabilities}

\item{n}{sample size}

\item{N}{population size}

\item{s}{vector of length N, with 1s at the positions of self-selecting units}

\item{list}{vector with positions of self selcting units}
}
\description{
Brewer sampling procedure --------------------------------------------------
}
\note{
this function is a modified version of function \code{\link[sampling]{UPbrewer}},
from the \pkg{sampling} package.
}
\keyword{internal}
