% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excludeSSU}
\alias{excludeSSU}
\title{Exclude self-selecting units}
\usage{
excludeSSU(pik, eps = 1e-06)
}
\arguments{
\item{pik}{vector of first-order inclusion probabilities}

\item{eps}{control value for pik}
}
\description{
Exclude self-selecting units and units with probability zero and returns a list
with parameters needed to perform sampling
}
\note{
the code is taken from package \pkg{sampling}
}
\keyword{internal}
