% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_approximations.R
\name{jip_Brewer}
\alias{jip_Brewer}
\title{Brewer's joint-inclusion probability approximations}
\usage{
jip_Brewer(pik, method)
}
\arguments{
\item{pik}{numeric, vector of first-order inclusion probabilities for all
population units.}

\item{method}{string representing one of the available approximation methods.}
}
\description{
Approximation of joint inclusion probabilities by one of the estimators
proposed by Brewer and Donadio (2003)
}
\details{
\code{"Brewer18"} is the approximation showed in equation (18) of Brewer and Donadio (2003)
}
