% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_approximations.R
\name{jip_Hajek}
\alias{jip_Hajek}
\title{Hájek's joint-inclusion probability approximation}
\usage{
jip_Hajek(pik)
}
\arguments{
\item{pik}{numeric, vector of first-order inclusion probabilities for all
population units.}
}
\description{
Estimate joint-inclusion probabilities using Hájek (1964) equation
}
