% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_approximations.R
\name{jip_HartleyRao}
\alias{jip_HartleyRao}
\title{Hartley-Rao approximation of joint-inclusion probabilities}
\usage{
jip_HartleyRao(pik)
}
\arguments{
\item{pik}{numeric, vector of first-order inclusion probabilities for all
population units.}
}
\description{
Approximation of joint-inclusion probabilities with precision of order
\eqn{O(N^{-4})} for the random systematic sampling design
by Hartley and Rao (1962), pag. 369 eq. 5.15
}
