% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_MonteCarlo.R
\name{jip_MonteCarlo}
\alias{jip_MonteCarlo}
\title{Approximate inclusion probabilities by Monte Carlo simulation}
\usage{
jip_MonteCarlo(x, n, replications = 1e+06, design, units, seed = NULL,
  asDataFrame = FALSE, writeOnFile = TRUE, file_path, by = 5e+05,
  spMat = NULL, beta = NULL)
}
\arguments{
\item{x}{size measure or first-order inclusion probabilities}

\item{n}{sample size (for fixed size samplings), or expected sample size (for Poisson sampling)}

\item{replications}{number of independent Monte Carlo replications}

\item{design}{sampling procedure to be used for sample selection.
Either a string indicating the name of the sampling design or a function;
see section "Details" for more information.}

\item{units}{id of units for which probabilities have to be estimated.
Optional, if missing, estimates are produced for the whole population}

\item{seed}{a valid seed value for reproducibility}

\item{asDataFrame}{logical, should output be in a data.frame form? if FALSE,
a matrix is returned}

\item{writeOnFile}{logical, should output be written on a text file?}

\item{file_path}{path of the directory where the text file should be created,
valid only if \code{writeOnFile} is TRUE}

\item{by}{integer scalar indicating every how many replications a partial output
should be saved}

\item{spMat}{a matrix of spatial contiguity among areas, only used for
FPDUST sampling}

\item{beta}{beta parameter to be used in FPDUST sampling (\code{beta<1})}

\item{design_pars}{only used when a function is passed to argument \code{design},
list of parameters to pass to the sampling design function.}
}
\value{
A matrix of estimated inclusion probabilities
}
\description{
Approximate first and second-order inclusion probabilities by means of
Monte Carlo simulation.
Estimates are obtained as proportion of the number of occurrences of each unit or
couple of units over the total number of replications. One unit is added to both
numerator and  denominator to assure strict positivity of estimates (Fattorini, 2006).
}
\details{
Argument \code{design} accepts either a string indicating the sampling design
to use to draw samples or a function.
Accepted designs are "brewer", "tille", "maxEntropy", "poisson",
"sampford", "systematic", "randomSystematic", "srs", chao", "sunter", "srs", "fpdust", "fpdust_pps".
The user may also pass a function as argument; such function should take as input
the parameters passed to argument \code{design_pars} and return either a logical
vector or a vector of 0 and 1,  where \code{TRUE} or \code{1} indicate sampled
units and \code{FALSE} or \code{0} indicate non-sample units.
The length of such vector must be equal to the length of \code{x}
if \code{units} is not specified, otherwise it must have the same length of \code{units}.
}
\examples{
### Generate population data ---
N <- 20; n<-5

set.seed(0)
x <- rgamma(N, scale=10, shape=5)
y <- abs( 2*x + 3.7*sqrt(x) * rnorm(N) )

pik  <- n * x/sum(x)

jipSim(x=pik, n = n, replications = 100, design = "brewer")

}
