% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jipMtoDF}
\alias{jipMtoDF}
\title{Transform matrix of Joint-Inclusion Probabilities to data.frame}
\usage{
jipMtoDF(jip, id = NULL)
}
\arguments{
\item{jip}{a square matrix of joint-inclusion probabilities, symmetric or
upper-triangular}

\item{id}{optional, vector of id labels, its length should be equal to
\code{ncol(jip)} and \code{nrow(jip)}}
}
\description{
Transform matrix of Joint-Inclusion Probabilities to data.frame
}
