% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEntropy.R
\name{maxEntropy}
\alias{maxEntropy}
\title{Conditional Poisson Sampling (maximum entropy sampling)}
\usage{
maxEntropy(pik, N, q)
}
\arguments{
\item{pik}{vector of first-order inclusion probabilities}

\item{N}{population size (excluding self-selecting units)}

\item{q}{matrix of selection probabilities, computed by means of function
\code{UPMEsfromq()} of package \code{sampling}.}
}
\description{
Draw a sample by means of Conditional Poisson Sampling
}
\note{
this functions is a modified version of function \code{\link[sampling]{UPmaxentropy}},
in the \pkg{sampling} package.
}
\keyword{internal}
