% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEntropy.R
\name{pre_CPS}
\alias{pre_CPS}
\title{Conditional Poisson Sampling  - compute selection probabilities}
\usage{
pre_CPS(pik)
}
\arguments{
\item{pik}{vector of first-order inclusion probabilities}
}
\description{
Compute matrix of selection probabilities for Conditional Poisson Sampling
}
\note{
this functions is a modified version of function \code{\link[sampling]{UPmaxentropy}},
in the \pkg{sampling} package.
}
