% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tille.R
\name{pre_tille}
\alias{pre_tille}
\title{Tillé's elimination procedure - elimination probabilities}
\usage{
pre_tille(pik)
}
\arguments{
\item{pik}{vector of first-order inclusion probabilities}
}
\description{
Computes a matrix with elimination probabilities for each step
of Tillé's elimination procedure
}
