% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampford.R
\name{sampford}
\alias{sampford}
\title{Rao-Sampford sampling}
\usage{
sampford(pik, n, N, s, list)
}
\arguments{
\item{pik}{vector of first-order inclusion probabilities}

\item{n}{sample size}

\item{N}{population size (excluding self-selecting units)}

\item{s}{vector of length N, with 1s at the positions of self-selecting units}

\item{list}{vector with positions of self selcting units}
}
\description{
Draw a sample by means of Rao-Sampford sampling
}
\note{
this function is a modified version of function \code{\link[sampling]{UPsampford}},
in the \pkg{sampling} package.
}
\keyword{internal}
