% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_output}
\alias{save_output}
\title{Save partial results}
\usage{
save_output(
  iteration,
  design_name,
  counts,
  units,
  filename,
  path,
  status,
  as_data_frame
)
}
\arguments{
\item{iteration}{integer indicating the iterations the simulation is at}

\item{design_name}{string indicating the name of the sampling design to include
in the filename}

\item{counts}{matrix with number of occurrences of couple of units up to
current replication}

\item{units}{id of units for which output should be saved}

\item{filename}{name of the file to write on disk}

\item{path}{path to the directory where the file should be saved}

\item{status}{\code{1} if partial results are written before the maximum number
of replications is reached, \code{0} otherwise}

\item{as_data_frame}{logical, should output be in form of a data frame?}
}
\description{
Write joint inclusion probability estimates on a file every \code{by} replications
}
\keyword{internal}
