% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{savepartial}
\alias{savepartial}
\title{Save partial results}
\usage{
savepartial(iteration, replications, by, counts, units, file_path, asDataFrame)
}
\arguments{
\item{iteration}{integer indicating the iterations the simulation is at}

\item{replications}{integer indicating the total number of replications
for the simulation}

\item{by}{integer indicating after how many iterations estimates should be
written on a file, must be < \code{replications}}

\item{counts}{matrix with number of occurrences of couple of units up to
current replication}

\item{units}{id of units for which output should be saved}

\item{file_path}{path of the file to write}

\item{asDataFrame}{logical, should output be in form of a data frame?}
}
\description{
Write joint inclusion probability estimates on a file every \code{by} replications
}
\keyword{internal}
